requirements.txt
flask
flask-sqlalchemy
SQLAlchemy>=2.0.36
Flask-Login==0.6.3
Werkzeug==3.0.1
psycopg2-binary>=2.9.10
pymysql==1.1.0
python-dotenv==1.0.1
requests==2.31.0
Flask-Caching==2.1.0
redis==5.0.1
gunicorn==21.2.0
reportlab==4.0.8
psutil==5.9.8
